requirements.txt
deepspeed==0.9.2
easydict==1.10
einops==0.6.1
ftfy==6.1.1
gradio==3.41.2
h5py==3.9.0
iopath==0.1.10
ipdb==0.13.13
kornia==0.7.0
matplotlib==3.7.2
mdtex2html==1.2.0
numpy==1.24.3
open3d_python==0.3.0.0
opencv_python==4.8.0.74
peft==0.3.0
Pillow==10.0.0
pytorchvideo==0.1.5
PyYAML==6.0.1
regex==2022.10.31
scikit_learn==1.3.0
termcolor==2.3.0
timm==0.6.7
#torch==1.13.1+cu117
#torchaudio==0.13.1
#torchvision==0.14.1
tqdm==4.64.1
transformers==4.29.1
sentencepiece
